refactor, move SemanticUtil test cases into new class
