+++ test/src/org/genepattern/server/dm/TestExternalFile.java
@@ -2,27 +2,10 @@ package org.genepattern.server.dm;
 
 import static org.junit.Assert.assertEquals;
 
-import org.genepattern.util.SemanticUtil;
 import org.junit.Test;
 
 public class TestExternalFile {
     public static final String externalHref="http://www.broadinstitute.org:8080";
-    
-    
-    protected void checkExtension(final String filename, final String expected_extension) {
-        assertEquals("getExtension("+filename+")", expected_extension,
-                SemanticUtil.getExtension(filename));
-    }
-    
-    @Test
-    public void semanticUtil_getExtensionFromName() {
-        checkExtension(null, null);
-        checkExtension("filename.txt", "txt");
-        checkExtension("name_no_ext", "");
-        checkExtension("dirname/", "");
-        checkExtension(".hidden.txt", "txt");
-        checkExtension(".hidden_no_ext", "");
-    }
     
     @Test
     public void fromUrlSpec_emptyPath() {
